# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Covenant API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/grunts": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetGrunts",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Grunt"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "CreateGrunt",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      },
      "put": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "EditGrunt",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      }
    },
    "/api/grunts/{id}": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetGrunt",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "DeleteGrunt",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/grunts/{name}": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetGruntByName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      }
    },
    "/api/grunts/guid/{guid}": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetGruntByGUID",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      }
    },
    "/api/grunts/originalguid/{serverguid}": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetGruntByOriginalServerGUID",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serverguid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      }
    },
    "/api/grunts/{id}/path/{cid}": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetPathToChildGrunt",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/grunts/{id}/outbound": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "GetOutboundGrunt",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grunt"
            }
          }
        }
      }
    },
    "/api/grunts/{id}/interact": {
      "post": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "InteractGrunt",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GruntCommand"
            }
          }
        }
      }
    },
    "/api/grunts/{id}/compileexecutor": {
      "get": {
        "tags": [
          "GruntApi"
        ],
        "operationId": "CompileGruntExecutor",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "byte",
              "type": "string"
            }
          }
        }
      }
    },
...
    "/api/commands/{id}": {
      "get": {
        "tags": [
          "GruntCommandApi"
        ],
        "operationId": "GetGruntCommand",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GruntCommand"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GruntCommandApi"
        ],
        "operationId": "DeleteGruntCommand",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "CommandOutput": {
      "required": [
        "gruntCommandId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "output": {
          "type": "string"
        },
        "gruntCommandId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Theme": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "backgroundColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "backgroundTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalBorderColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "navbarColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "sidebarColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputDisabledColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "textLinksColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "codeMirrorTheme": {
          "$ref": "#/definitions/CodeMirrorTheme"
        }
      }
    },
    "CovenantUser": {
      "type": "object",
      "properties": {
        "themeId": {
          "format": "int32",
          "type": "integer"
        },
        "theme": {
          "$ref": "#/definitions/Theme"
        },
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "ImplantLanguage": {
      "enum": [
        "CSharp"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantLanguage",
        "modelAsString": false
      }
    },
    "CommunicationType": {
      "enum": [
        "HTTP",
        "SMB",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "CommunicationType",
        "modelAsString": false
      }
    },
    "ImplantDirection": {
      "enum": [
        "Push",
        "Pull"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantDirection",
        "modelAsString": false
      }
    },
    "ListenerType": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "DotNetVersion": {
      "enum": [
        "Net35",
        "Net45",
        "NetCore31"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "DotNetVersion",
        "modelAsString": false
      }
    },
    "ImplantTemplate": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "commType": {
          "$ref": "#/definitions/CommunicationType"
        },
        "implantDirection": {
          "$ref": "#/definitions/ImplantDirection"
        },
        "compatibleListenerTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListenerType"
          },
          "readOnly": true
        },
        "compatibleDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "stagerCode": {
          "type": "string"
        },
        "executorCode": {
          "type": "string"
        }
      }
    },
...
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/ProfileType"
        },
        "messageTransform": {
          "type": "string"
        }
      }
    },
    "ListenerStatus": {
      "enum": [
        "Uninitialized",
        "Active",
        "Stopped"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ListenerStatus",
        "modelAsString": false
      }
    },
    "Listener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "guid",
        "listenerTypeId",
        "name",
        "profileId",
        "status"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "guid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantUrl": {
          "type": "string"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "RuntimeIdentifier": {
      "enum": [
        "win_x64",
        "win_x86",
        "win_arm",
        "win_arm64",
        "win7_x64",
        "win7_x86",
        "win81_x64",
        "win81_x86",
        "win81_arm",
        "win10_x64",
        "win10_x86",
        "win10_arm",
        "win10_arm64",
        "linux_x64",
        "linux_musl_x64",
        "linux_arm",
        "linux_arm64",
        "rhel_x64",
        "rhel_6_x64",
        "tizen",
        "tizen_4_0_0",
        "tizen_5_0_0",
        "osx_x64",
        "osx_10_10_x64",
        "osx_10_11_x64",
        "osx_10_12_x64",
        "osx_10_13_x64",
        "osx_10_14_x64",
        "osx_10_15_x64"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "RuntimeIdentifier",
        "modelAsString": false
      }
    },
    "GruntStatus": {
      "enum": [
        "Uninitialized",
        "Stage0",
        "Stage1",
        "Stage2",
        "Active",
        "Lost",
        "Exited",
        "Disconnected",
        "Hidden"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "GruntStatus",
        "modelAsString": false
      }
    },
    "IntegrityLevel": {
      "enum": [
        "Untrusted",
        "Low",
        "Medium",
        "High",
        "System"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "IntegrityLevel",
        "modelAsString": false
      }
    },
...
    "GruntTaskingType": {
      "enum": [
        "Assembly",
        "SetDelay",
        "SetJitter",
        "SetConnectAttempts",
        "SetKillDate",
        "Exit",
        "Connect",
        "Disconnect",
        "Tasks",
        "TaskKill"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "GruntTaskingType",
        "modelAsString": false
      }
    },
...
    "GruntTasking": {
      "required": [
        "gruntId",
        "gruntTaskId",
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "gruntId": {
          "format": "int32",
          "type": "integer"
        },
        "grunt": {
          "$ref": "#/definitions/Grunt"
        },
        "gruntTaskId": {
          "format": "int32",
          "type": "integer"
        },
        "gruntTask": {
          "$ref": "#/definitions/GruntTask"
        },
        "type": {
          "$ref": "#/definitions/GruntTaskingType"
        },
        "parameters": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "$ref": "#/definitions/GruntTaskingStatus"
        },
        "taskingTime": {
          "format": "date-time",
          "type": "string"
        },
        "completionTime": {
          "format": "date-time",
          "type": "string"
        },
        "gruntCommandId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "GruntCommand": {
      "required": [
        "command",
        "commandOutputId",
        "commandTime",
        "userId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "command": {
          "type": "string"
        },
        "commandTime": {
          "format": "date-time",
          "type": "string"
        },
        "commandOutputId": {
          "format": "int32",
          "type": "integer"
        },
        "commandOutput": {
          "$ref": "#/definitions/CommandOutput"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/CovenantUser"
        },
        "gruntTaskingId": {
          "format": "int32",
          "type": "integer"
        },
        "gruntTasking": {
          "$ref": "#/definitions/GruntTasking"
        },
        "gruntId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Grunt": {
      "required": [
        "connectAttempts",
        "delay",
        "dotNetVersion",
        "implantTemplateId",
        "integrity",
        "jitterPercent",
        "killDate",
        "name",
        "originalServerGuid",
        "runtimeIdentifier",
        "smbPipeName",
        "status",
        "useCertPinning",
        "validateCert"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "originalServerGuid": {
          "type": "string"
        },
        "guid": {
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "implantTemplateId": {
          "format": "int32",
          "type": "integer"
        },
        "implantTemplate": {
          "$ref": "#/definitions/ImplantTemplate"
        },
        "validateCert": {
          "type": "boolean"
        },
        "useCertPinning": {
          "type": "boolean"
        },
        "smbPipeName": {
          "type": "string"
        },
        "listenerId": {
          "format": "int32",
          "type": "integer"
        },
        "listener": {
          "$ref": "#/definitions/Listener"
        },
        "note": {
          "type": "string"
        },
        "delay": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 0,
          "type": "integer"
        },
        "jitterPercent": {
          "format": "int32",
          "maximum": 100,
          "minimum": 0,
          "type": "integer"
        },
        "connectAttempts": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 0,
          "type": "integer"
        },
        "killDate": {
          "format": "date-time",
          "type": "string"
        },
        "dotNetVersion": {
          "$ref": "#/definitions/DotNetVersion"
        },
        "runtimeIdentifier": {
          "$ref": "#/definitions/RuntimeIdentifier"
        },
        "status": {
          "$ref": "#/definitions/GruntStatus"
        },
        "integrity": {
          "$ref": "#/definitions/IntegrityLevel"
        },
        "process": {
          "type": "string"
        },
        "userDomainName": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "ipAddress": {
          "type": "string"
        },
        "hostname": {
          "type": "string"
        },
        "operatingSystem": {
          "type": "string"
        },
        "gruntSharedSecretPassword": {
          "type": "string"
        },
        "gruntRSAPublicKey": {
          "type": "string"
        },
        "gruntNegotiatedSessionKey": {
          "type": "string"
        },
        "gruntChallenge": {
          "type": "string"
        },
        "activationTime": {
          "format": "date-time",
          "type": "string"
        },
        "lastCheckIn": {
          "format": "date-time",
          "type": "string"
        },
        "powerShellImport": {
          "type": "string"
        },
        "gruntCommands": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GruntCommand"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}